Final Year Project
Final Year Project:  Distributed System: Tickets on-line sale
This project surveys and researches in the area of distributed systems, it uses network socket to design and create a system that is used to book tickets. Users could use this system to book train tickets and coach tickets.
Implement this system to manage the communication between the server and multiple clients by using Java Sockets and Threads.
Technology: Java, SQL.